+++ packages/sources/cryptoapis-v2/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @chainlink/cryptoapis-v2-adapter
 
+## 1.1.20
+
+### Patch Changes
+
+- 8d6ff4693: Added buildUrl & buildUrlPath methods to util. Updated source adapters to use these methods for building URLs with user input.
+- Updated dependencies [8d6ff4693]
+- Updated dependencies [196336176]
+  - @chainlink/ea-bootstrap@1.13.0
+  - @chainlink/ea-factories@1.0.34
+  - @chainlink/ea-test-helpers@1.2.0
+
+## 1.1.19
+
+### Patch Changes
+
+- Updated dependencies [effb61e40]
+  - @chainlink/ea-bootstrap@1.12.2
+  - @chainlink/ea-factories@1.0.33
+  - @chainlink/ea-test-helpers@1.2.0
+
 ## 1.1.18
 
 ### Patch Changes