+++ packages/sources/jpegd/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @chainlink/jpegd-adapter
 
+## 1.1.2
+
+### Patch Changes
+
+- Updated dependencies [8d6ff4693]
+- Updated dependencies [196336176]
+  - @chainlink/ea-bootstrap@1.13.0
+  - @chainlink/ea-test-helpers@1.2.0
+
+## 1.1.1
+
+### Patch Changes
+
+- Updated dependencies [effb61e40]
+  - @chainlink/ea-bootstrap@1.12.2
+  - @chainlink/ea-test-helpers@1.2.0
+
 ## 1.1.0
 
 ### Minor Changes